requirements.txt
karton-core==5.*
